predictor/3D_GNN/requirements.txt
ase==3.22.1
easydict==1.9
matplotlib==3.6.2
networkx==2.8.6
numpy==1.22.4
pandas==1.5.0
PyYAML==6.0.2
rdkit_pypi==2022.9.5
scikit_learn==1.1.1
scipy==1.7.1
torch_cluster==1.6.0
torch_geometric==2.0.4
torch_scatter==2.0.9
torch_sparse==0.6.14
tqdm==4.64.1
